% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{Capture location}
  \item{MILE}{Distance from the mouth of the river or creek}
  \item{SPECIES}{Species of fish}
  \item{LENGTH}{Length of fish}
  \item{WEIGHT}{Weight of fish}
  \item{DDT}{DDT levels}
  ...
}
}
\source{
<https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
ddt
}
\description{
A complete data set of measurements of 144 fish caught in various rivers
}
\keyword{datasets}
